Optimization

+++ download.sh
@@ -5,8 +5,12 @@ archive_filename="AT&T UNIX System V Release 4 Version 2.1 (3.5).7z"
 intel_archive_url="https://archive.org/download/intel-unix-sv-r4-v2/Intel%20Unix%20SVR4V2%20%285.25%20Floppy%29.zip"
 intel_archive_name="Intel Unix SVR4V2 (5.25 Floppy)"
 
-curl -L $archive_url -o "$archive_filename"
-7z x "$archive_filename"
+if [ ! -f "$archive_filename" ]; then
+        curl -L $archive_url -o "$archive_filename"
+        7z x "$archive_filename"
+fi
 
-curl -L $intel_archive_url -o "$intel_archive_name.zip"
-unzip "$intel_archive_name.zip" -d "$intel_archive_name"
+if [ ! -f "$intel_archive_name.zip" ]; then
+        curl -L $intel_archive_url -o "$intel_archive_name.zip"
+        unzip "$intel_archive_name.zip" -d "$intel_archive_name"
+fi
